API structure with first champion API
The commit lays out the basic structure for API calls including a base
model, a request model, and re allocation of semantics from client to
request model. Client is no longer responsible for making requests,
requests are responsible for making the request. Client holds settings
pertaining to a group of api calls the user may want to make, and
provides the user access to that api.

Specs included